pyproject.toml
[project]
name = "h1-network-universe-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adbc-driver-postgresql>=0.4.0",
    "boto3>=1.34.0",
    "geoalchemy2>=0.14.0",
    "geopandas>=1.0.0",
    "openpyxl>=3.1.5",
    "polars>=1.38.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pyogrio>=0.7.0",
    "sqlalchemy>=2.0.48",
]

[project.optional-dependencies]
dev = [
    "moto[sqs,s3]>=5.0.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
]

[project.scripts]
pipeline-worker = "pipeline.worker.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pipeline"]
